pyproject.toml
[project]
name = "gpt"
version = "0.1.0"
description = "like nanogpt, but not really"
readme = "README.md"
authors = [
    { name = "k", email = "keatone@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "torchtitan",
    "ruff>=0.14.3",
    "loguru>=0.7.3",
    "triton>=3.5.0",
    "pytorch-triton",
    "torch>=2.10.0.dev20251029",
    "torchao>=0.15.0.dev20251030",
    "flash-linear-attention>=0.4.0",
]

[build-system]
requires = ["uv_build>=0.9.6,<0.10.0"]
build-backend = "uv_build"

[project.scripts]
train = "gpt:main"

[tool.uv]
prerelease = "allow"

[tool.uv.build-backend]
module-name = "gpt"
module-root = ""

[tool.uv.sources]
torch = { index = "pytorch-cu130" }
torchao = { index = "pytorch-cu130" }
pytorch-triton = { index = "pytorch-cu130" }
torchtitan = { git = "https://github.com/pytorch/torchtitan.git" }

[[tool.uv.index]]
name = "pytorch-cu130"
url = "https://download.pytorch.org/whl/nightly/cu130"
explicit = true

[tool.ruff]
line-length = 120

[tool.ruff.lint]
extend-select = [
    "F",        # Pyflakes rules
    "W",        # PyCodeStyle warnings
    "E",        # PyCodeStyle errors
    "I",        # Sort imports properly
    "UP",       # Warn if certain things can changed due to newer Python versions
    "C4",       # Catch incorrect use of comprehensions, dict, list, etc
    "FA",       # Enforce from __future__ import annotations
    "ISC",      # Good use of string concatenation
    "ICN",      # Use common import conventions
    "RET",      # Good return practices
    "SIM",      # Common simplification rules
    "TID",      # Some good import practices
    "TC",       # Enforce importing certain types in a TYPE_CHECKING block
    "PTH",      # Use pathlib instead of os.path
    "TD",       # Be diligent with TODO comments
    "NPY",      # Some numpy-specific things
    "COM",      # enforce trailing comma rules
    "DTZ",      # require strict timezone manipulation with datetime
]
